+++ s/core/testing/setup-math-spy.ts
@@ -1,5 +1,5 @@
 
-import {fns} from "../types.js"
+import {asFns} from "../types.js"
 
 export type MathFns = {
 	add(a: number, b: number): Promise<number>
@@ -12,7 +12,7 @@ export function setupMathSpy() {
 		mul: [] as {a: number, b: number, ret: number}[],
 	}
 
-	const mathFns = fns<MathFns>({
+	const mathFns = asFns<MathFns>({
 		async add(a: number, b: number) {
 			const ret = a + b
 			calls.add.push({a, b, ret})